finagle-core: integrate and expose Dtab.limited
Problem

We want to introduce a limited Dtab that does not
propagate via request broadcast and stays local
to the process. We need to make sure that's integrated
where Dtab.local is used and assert that we are not
leaking this information out along with the Dtab.local.

Solution

Expose the `Dtab.limited` API publicly, integrate with the
`Namer` and `BindingFactory`, as well as test the expected
behavior. We expect that `Dtab.local`s that are set for a
request will take precedence over the `Dtab.limited`, but
both will override values set in the `Dtab.base`.

JIRA Issues: CSL-10977

Differential Revision: https://phabricator.twitter.biz/D677860