metrics: Remove MetricSchema and make it an attribute of MetricBuilder
Problem

The metric types are defined in a wrapper class MetricSchema - CounterSchema,
GaugeSchema, and HistogramSchema. This is becoming inconvenient that we need to
re-wrap MetricBuilders when we made a change to them. And requires StatsReceiver
to carry over the metric type info, while ideally, it should stick with
MetricBuilder itself.

Solution

Remove the wrapper and add enum fields in MetricBuilder.

JIRA Issues: CSL-10923

Differential Revision: https://phabricator.twitter.biz/D668739