finatra: Add scoverage excludes
Problem/Solution

Some deprecated code with replacements elsewhere are being included in
code coverage metrics. Lets add them to the scoverage excludes to not be
included in the generate coverage metrics. We also disable coverage reporting
for the example applications and update the finatra/benchmarks/build.sbt
formatting for the new "slash" style expected with sbt 1.5.x.

Differential Revision: https://phabricator.twitter.biz/D684674