finatra/jackson: fix Annotations cache leak
Problem

The `Annotations` cache is leaking resources.
`BeanProperty` does not implement `hashCode` and
`equals` method, which means that we are generating
a new cache entry on every call and retaining the entry
in the cache indefinitely, which will eventually result in
an `OutOfMemory` (or GC spiral). We aren't gaining
any performance improvement from the cache due to
this.

Solution

Let's remove the cache and resolve the leak.

JIRA Issues: CSL-11047

Differential Revision: https://phabricator.twitter.biz/D688123