inject-core: TestMixinTest#assertFailedFuture should always fail when passed an unfailed future
Problem

TestMixin#assertFailedFuture will pass if it's passed an unfailed future but the
type of exception we're testing for is TestFailedFuture or one of its supertypes.
Basically the problem is that we call `fail` in a try/catch block and we sometimes
catch `TestFailedFuture` without handling it delicately.

Solution

ScalaTest's intercept already has the behavior that we want, so just rely on that.

JIRA Issues: CSL-11174

Differential Revision: https://phabricator.twitter.biz/D705002