Revert "Twitter-oss: update OSS libraries post-release to 21.9.0"
Problem:

Sonatype requires a javadoc jar, even if empty to release.

Solution / Result:

Revert the post-release commit to act as the new SHA to which
will be released. This SHA will come after D724170 which doesn't
skip the creation of the empty util javadoc JAR for Scala 3.

JIRA Issues: CSL-11162

Differential Revision: https://phabricator.twitter.biz/D724302