finatra: Unflaky inject-logback tests on JDK11
Problem

Somehow stopping the AsyncAppender's worker thread right after it started can leave the underlying
queue in an undefined state (the lock is never released by the thread that abruptly terminated).

Solution

One way to prevent the worker thread from doing its work is to substitute it with one that does
nothing. This way we don't need to stop it.

JIRA Issues: CSL-11276

Differential Revision: https://phabricator.twitter.biz/D735655