finatra/http-server: Better `HttpRouter` Java support for adding generically typed Filters
Problem

We do not properly document how Java users can apply generically typed
Filters to the `HttpRouter`. Which right now requires Java users that want
the injector to do the heavy lifting of Filter construction to manually obtain
an instance of the Filter from the object graph to apply it to the `HttpRouter`
(or play games with Scala type erasure from Java to create Manifests to call
the `HttpRouter#filter[T: Manifest]` function with enough type information).

The Guice recommended way to bind parameterized types is via usage
of a `TypeLiteral`.

Solution

Add versions of the `HttpRouter#filter` method which takes in a Guice
TypeLiteral which can help Java users properly use generically typed
Filters with the `HttpRouter`.

JIRA Issues: CSL-11419

Differential Revision: https://phabricator.twitter.biz/D768777