finagle-core: Store maxEffortExhausted counter as a val in Balancer
Problem:
There used to be a java.lang.AbstractMethodError scalac compile/runtime bug. As
a workaround, we had to store the maxEffortExhausted counter as an abstract
method. The bug is fixed now.

Solution:
Store maxEffortExhausted as a member val on the Balancer trait. This simplifies
the code because the counter only needs to be defined once in the trait. Remove
overridden maxEffortExhausted counters in load balancer implementations that
mix in or extend Balancer.

Differential Revision: https://phabricator.twitter.biz/D775764