inject-core: add 'runAfterAll' hook to IntegrationTestMixin
Problem

The `EmbeddedTwitterServer` has a `closeOnExit` method for
registering logical hooks for cleaning up test resources
when the server shuts down, but we do not have a general
utility for registering shutdown logic that is detached from
the `EmbeddedTwitterServer` lifecycle *in a simple, consistent
manner*. The current solution is to override ScalaTest
`afterAll` directly, which comes with its own pitfalls. This
can result in `super.afterAll` not being called by an inheritor
and not handling exceptions as part of clean-up can result in
unexpected TestSuite failures.

Solution

Introduce a `runAfterAll` method that can be called to register
shutdown logic that will not fail the `TestSuite` as a result
of an exception during clean-up. It should be highlighted that
we do not prevent anyone from explicitly overriding `afterAll`
for logic that should result in a test failure as a result of
an exception.

JIRA Issues: CSL-11136

Differential Revision: https://phabricator.twitter.biz/D707939