finatra: treat non-constant routes with trailing slash as a different route
Problem
We don't allow users to register a non-constatn route
that has the same segments but with a different trailing
slash variant. For example, registering the following routes
to the same controller is not allowed:

```
"/user/:id"
"/user/:id/"
```

Solution
We should be able to treat routes with trailing slash as a
different routes, this can be achieved by mutating the key
Finatra used as path segment, to allow `/` as part of the
key if it appears to be the last character of a path.

JIRA Issues: CSL-11479

Differential Revision: https://phabricator.twitter.biz/D787560