finatra: Create service before checking for service class name in AbstractThriftServer
Problem:
The ServiceClazzStackParam is None before a Java based ThriftServer service is
created and configured, causing the Java based ThriftMux server to be created
without a service class name.

Solution:
Create and configure the service in AbstractThriftServer first before calling
`getServiceClazzStackParam` to set the service class name. When
JavaThriftRouter adds the controller, ServiceCreator#apply creates the service
and defines the ServiceClazzStackParam with the class name of the controller.

JIRA Issues: FTEST-23

Differential Revision: https://phabricator.twitter.biz/D793335