finatra/mysql-client: introduce project
Problem

We have the `finagle-mysql` project, but no standardized way of
using a `TwitterModule` to inject a Finagle client within Finatra's
framework. This leads to multiple implementations of the same
basic functionality and no consistent utility for testing this behavior.
This is extra difficult with `finagle-mysql` because the type bindings
require use of mixin traits (i.e. `Client with Transactions`).

Solution

Let's introduce a `finatra/mysql-client` project which makes it easy
to bind a `finagle-mysql` client and a `EmbeddedMysqlServer` that
wraps our `finagle-mysql` test harness utilities to test against a real
MySQL instance and allow for easily binding over our client module
to use the `EmbeddedMysqlServer` within `FeatureTest`s.

We also have a `TestLifecycle` trait for managing the lifecycle of an
`EmbeddedMysqlServer` in the same way a `FeatureTest` does for
an `EmbeddedTwitterServer`. As a result, we are moving the
`runAfterAll` hook into the more generic `TestMixin`

Differential Revision: https://phabricator.twitter.biz/D784576