finatra/http-server: Allow differing HTTP/HTTPS listening server configuration
Problem

While we expose serving either an HTTP or an HTTPS ListeningServer (or both) we
do not allow for a way to customize the Finagle `Service[-R,+R]` that is served
on either. Currently both are only able to serve the external service created
by the `HttpRouter`. However, there are occasions where users may want the services
to differ.

Additionally, there is not a clean way to run both the HTTP and the HTTPS listening
servers together in tests (it requires hacking one or the other in the server under
test to run on the ephemeral port and letting the testing infrastructure set the other
via the flag).

Solution

Since the `HttpRouter` is a stateful singleton, introducing a fully separate construct
or making the current router not a singleton is perhaps bit too far though something
to consider in the future (there should perhaps be an analogous HttpsServer trait with
an HttpsRouter and all of the logic to configure and serve the Https ListeningServer
moved there). Instead, we break up the current `build` method which creates a ListeningServer
from a given address and Finagle `Http.Server` into two versions which can be overridden
by users to re-specify how the ListeningServer is created, which essentially allows for
specifying how the Finagle `Service[-R, +R]` is created and served by the `Http.Server`.

We also update the `EmbeddedHttpServer` and `EmbeddedHttpClient` to allow for being able
to run both the HTTP and HTTPS listening servers in tests. This is done by setting the
`httpsPortFlag` to the value of `https.port` which will enable the binding of the HTTPS listening
server to the ephemeral port in tests.

Result

This allows for users to specify divergent logic to be served by the Http or Https
ListeningServer. Note these changes don't affect the creation of the ListeningServers.
It remains true that if the respective port is not set, the ListeningServer will not
be created. e.g., if an https port is not given, the Https ListeningServer will not
be created.

Lastly, users can run both the HTTP and the HTTPS listening servers easily in tests by
setting the `httpsPortFlag` to `Some("https.port")`.

JIRA Issues: CSL-11584

Differential Revision: https://phabricator.twitter.biz/D809633