3rdparty: Revert "Revert "3rdparty: jackson upgrade to 2.13.0"" & Upgrade to 2.13.1
Problem

We're on an old version of jackson, and we'd like to upgrade to a new version.
We tried in the past to upgrade to 2.13.0, but it introduced a regression, where
they were double-serializing.  We believe 2.13.1 no longer has that regression,
so it should be safe to upgrade.

Solution

This reverts commit e27353b784543196b9f5fd572110e4b4fe04f4c4.

JIRA Issues: CSL-11573

Differential Revision: https://phabricator.twitter.biz/D808049