finatra-inject: update flaky noninjection tests
Problem:
We're seeing the following errors in a couple of
our finatra inject tests:

```
[info] - TwitterServer#stats receivers *** FAILED ***
Map() was empty (FeatureTestNonInjectionCustomStatsReceiverTest.scala:44)
```
```
[info] - server#in memory stats receiver with non-injectable server *** FAILED ***
[info]   Map() was empty (EmbeddedTwitterServerIntegrationTest.scala:237)
```

These tests are founded on the assumption that the gauges are
always updated with the finagle build revision metric (`build/revision`),
but that gauge is only added for injected twitter servers. both of these tests
are for the non-injectable server case.

Solution:
Lets instead just check that the offered stats receiver is the one used by
the server, since there's no gauge that is consistently in statsReceiver.gauges map.

JIRA Issues: CSL-11756

Differential Revision: https://phabricator.twitter.biz/D839829