inject-app: introduce test-friendly ConsoleWriter
Problem

When writing a `c.t.inject.App` and using the `EmbeddedApp`
to test, there is no common way to access the resulting
output of the `EmbeddedApp.main()` call. This is especially
important for command-line utilities that require a strict
output and/or expect that output to be piped to another
command.

Solution

Introduce `ConsoleWriter`, `ConsoleWriterModule`, and
`TestConsoleWriter`. The `ConsoleWriter` gets bound to a
default `System.out` and `System.err`, as well as scoping
the underlying App's scala `Console`. This allows binding
of the `TestConsoleWriter` which will record output and
give access to the `out` and `err` of the App for inspection
during tests. This is also done in a way to ensure that the
JVM global `out` and `err` streams are not modified to prevent
the inspected streams from being corrupted, resulting in flaky
tests.

Result

We now have a common way for accessing command-line output without
requiring every tool owner to implement their own solution.

JIRA Issues: CSL-11831

Differential Revision: https://phabricator.twitter.biz/D856159