inject-server: use NullStatsReceiver instead of InMemoryStatsReceiver in EmbeddedHttpClient
Problem/Solution

The `c.t.inject.server.EmbeddedHttpClient` installs an `InMemoryStatsReceiver`, but the
stats are never exposed or accessible anywhere. For large test suites or execution that
does not frequently clear the stats, this adds latency and essentially leaks the Histogram
stats data. As this provides no use and the `c.t.finatra.thrift.ThriftClient` uses a
`NullStatsReceiver`, let's make the change here to also use a `NullStatsReceiver` and trim
out the extra overhead.

JIRA Issues: CSL-11844

Differential Revision: https://phabricator.twitter.biz/D859449