inject-app: Remove SLF4J-API logging bridges as dependencies
Problem

It is incorrect to blindly specify the SLF4J-API logging bridges
on the inject-app library of the framework at it causes trouble
for usages that cannot use Logback as an SLF4J-API logging
implementation.

Solution

Remove the SLF4J-API logging bridges as dependencies. Users
should provide the correct bridges as needed for their chosen
SLF4J-API logging implementation. Note that the `TwitterServer`
logging libraries will bring the correct bridges for the
given logging implementation so for users that include a
dependency on one of those libraries, there is nothing else that
needs to be done.

Result

Less conflicting logging implementations.

JIRA Issues: CSL-9669

Differential Revision: https://phabricator.twitter.biz/D854393