finatra: introduce InMemoryTracer for inspecting Trace information
Problem

There is no standardized convention for verifying trace annotations
for a given service. While the `c.t.finagle.tracing.BufferingTracer`
exists, the utility is limited and there is no easy way to bind or
access the recorded trace information.

Solution

Introduce an `InMemoryTracer` that can be used to record trace
information and interrogated within a test to verify the presence
of annotations. The `StackClientModuleTrait` and server implementations
are modified to configure the `Tracer` on the respective client or server
to use the `Tracer` that is *optionally* bound to the injector, otherwise uses
the default tracer configuration. This results in no behavior change to a
runtime service outside of tests, as it will continue to bind the
`DefaultTracer` and allow for overriding configuration for clients and
servers that already set a non-default `Tracer` configuration. The
`EmbeddedTwitterServer` and its variants hook into this new integration
and bind an `InMemoryTracer`. The trace information is by default reset
between each test case in a `FeatureTest`.

Result

Using an `EmbeddedTwitterServer` or its variants will by default bind
an `InMemoryTracer` and expose access to its utilities for testing.

The `InMemoryTracer` also supports dumping out Zipkin formatting JSON,
which can be uploaded to a Zipkin UI to verify the presence and formatting
of trace data.

Verifying the presence of trace data, especially for trace information that
may cross the Finagle thread context (i.e. 3rdparty libraries), is a much
faster iteration loop. Without this data, it requires running or deploying
a service, triggering a request with a sampled trace, and finding the generated
trace span in a UI, as well as searching through the UI for the correct information.
We can now automate this and verify in the Zipkin UI that our trace information
is well formed and well formatted.

Differential Revision: https://phabricator.twitter.biz/D878616