inject-server: Remove DeprecatedLogging trait
Problem

We added a shim to keep the flags defined by the util-logging (JUL) Logging
trait defined such that not all Finatra servers immediately died because
users we setting values for these flags even though they use Logback.

Reminder: by default setting the value for a flag that is not defined in your
server causes your server to fail startup with an exception and a message
which prints all the available flags to configure.

Solution

We're finally removing this shim. Users need to stop setting the util-logging
JUL flags since they do nothing.

Differential Revision: https://phabricator.twitter.biz/D850852