inject-utils: Move package private methods from PossiblyRetryable to ExceptionUtils
Problem

The package private methods `PossiblyRetryable#isCancellation` and
`PossibleRetryable#isNonRetryable` in inject-thrift are generally useful
for processing interrupts on Futures. Making them publicly available
allows users to not have to re-implement this logic (e.g., in the case
they are attempting to mask cancellations on a Future).

Solution

Move the package private methods in `PossibleRetryable` in inject/inject-thrift
to public methods in `ExceptionUtils` in inject/inject-utils.

Result

Less need for users to do this themselves (or hack around the visibility
of these methods to call them).

Differential Revision: https://phabricator.twitter.biz/D935287