finatra: Remove legacy DiffyModule
Problem/Solution

The DarkTrafficFilterModule that requires ServiceIface and ServiceIfaceBuilder
was deprecated in 2019. It's time to let it go.

ReqRepDarkTrafficFilterModule is pretty standalone that makes itself not
worrying about dependency injection, so unlikely we will see binding failures if
we only migrate DiffyModule.

However, essentially the Module will be chained in the server filter chain, if
the server still uses the legacy controller, which forces the Req/Rep type, we
can see requests cast failing on the run from Arg to scrooge.Request. Note that
this will only happen when it starts to send dark traffic, not during startup.

Solution

Migrate the services using legacy DiffyModule entirely new to the new finatra
ecosystem.

Differential Revision: https://phabricator.twitter.biz/D986488